Split OpenAIClient from jira_client.py
