+++ src/mantis/openai_client.py
@@ -1,0 +1,31 @@
+from typing import TYPE_CHECKING
+
+from openai import OpenAI
+
+
+if TYPE_CHECKING:
+    from mantis.jira.jira_client import JiraClient
+    from mantis.mantis_client import MantisClient
+
+
+class OpenAIClient:
+    def __init__(self, mantis: 'MantisClient', jira_client: 'JiraClient') -> None:
+        self.mantis = mantis
+        self.jira_client = jira_client
+        self.disabled = not self.mantis.options.chat_gpt_activated
+        self.client = OpenAI(base_url=self.mantis.options.chat_gpt_base_url, api_key=self.mantis.options.chat_gpt_api_key)
+        
+    def get_completion(self, input_text: str, prompt: str, model: str = 'gpt-4.1') -> str:
+        if self.disabled:
+            raise ConnectionError('OpenAI connectivity has not been configured')
+        completion = self.client.chat.completions.create(
+                model="gpt-4.1",
+                messages=[
+                    {"role": "developer", "content": prompt},
+                    {"role": "user", "content": input_text}
+                ]
+            )
+        response = completion.choices[0].message.content
+        if not response:
+            raise ValueError("Response is empty")
+        return response
